refactor: プリセットへモデルモジュールを追加
